+++ src/main/kotlin/us/frollo/frollosdk/extensions/ModelExtension.kt
@@ -184,6 +184,7 @@ private fun appendTransactionFilterToSqlQuery(sqlQueryBuilder: SimpleSQLiteQuery
     filter.accountIds?.let { if (it.isNotEmpty()) sqlQueryBuilder.appendSelection(selection = "t.account_id IN (${ it.joinToString(",") })") }
     filter.merchantIds?.let { if (it.isNotEmpty()) sqlQueryBuilder.appendSelection(selection = "t.merchant_id IN (${ it.joinToString(",") })") }
     filter.transactionCategoryIds?.let { if (it.isNotEmpty()) sqlQueryBuilder.appendSelection(selection = "t.category_id IN (${ it.joinToString(",") })") }
+    filter.billId?.let { sqlQueryBuilder.appendSelection(selection = "t.bill_id = $it") }
     filter.budgetCategory?.let { sqlQueryBuilder.appendSelection(selection = "t.budget_category = '${ it.name }'") }
     filter.baseType?.let { sqlQueryBuilder.appendSelection(selection = "t.base_type = '${ it.name }'") }
     filter.status?.let { sqlQueryBuilder.appendSelection(selection = "t.status = '${ it.name}'") }